+++ lua/render-markdown/init.lua
@@ -224,6 +224,7 @@ local M = {}
 ---@field public below? integer
 
 ---@class (exact) render.md.UserConfigOverrides
+---@field public buflisted? table<boolean, render.md.UserBufferConfig>
 ---@field public buftype? table<string, render.md.UserBufferConfig>
 ---@field public filetype? table<string, render.md.UserBufferConfig>
 
@@ -790,14 +791,16 @@ M.default_config = {
     --   heading, paragraph, code, dash, bullet, checkbox, quote, pipe_table,
     --   callout, link, sign, indent, latex, html, win_options
     overrides = {
-        -- Overrides for different buftypes, see :h 'buftype'
+        -- Override for different buflisted values, see :h 'buflisted'
+        buflisted = {},
+        -- Override for different buftype values, see :h 'buftype'
         buftype = {
             nofile = {
                 padding = { highlight = 'NormalFloat' },
                 sign = { enabled = false },
             },
         },
-        -- Overrides for different filetypes, see :h 'filetype'
+        -- Override for different filetype values, see :h 'filetype'
         filetype = {},
     },
     -- Mapping from treesitter language to user defined handlers